LatestConfigs/data-processing/requirements.txt
kafka-python
nltk